+++ src/main/java/me/tye/filemanager/FileGui.java
@@ -9,6 +9,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.inventory.InventoryClickEvent;
+import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -19,6 +20,7 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
@@ -33,12 +35,18 @@ public class FileGui implements Listener {
     public static HashMap<String, PathHolder> position = new HashMap<>();
 
     public static void openFolder(Player player) {
+        if (!player.hasPermission("fileman.file.nav")) {
+            fileData.remove(player.getUniqueId());
+            position.remove(player.getName());
+            return;
+        }
+
         Inventory gui = Bukkit.createInventory(player, 54, ChatColor.BLUE+"~"+position.get(player.getName()).getRelativePath()+ChatColor.GOLD+" $");
         List<Path> paths;
         try {
             paths = Files.list(Path.of(position.get(player.getName()).getCurrentPath())).toList();
         } catch (Exception e) {
-            log(e, player, Level.SEVERE, "There was an error trying to get the files in that folder.");
+            log(e, player, Level.WARNING, "There was an error trying to get the files in that folder.");
             return;
         }
 
@@ -69,6 +77,8 @@ public class FileGui implements Listener {
                 content.add(itemProperties(new ItemStack(Material.ARROW), "Up", List.of("Goes up a file.")));
             } else if (i == 4) {
                 content.add(itemProperties(new ItemStack(Material.BARRIER), "Exit", List.of("Closes the gui.")));
+            } else if (i == 8) {
+                content.add(itemProperties(new ItemStack(Material.GREEN_CONCRETE), "Create file", List.of("Creates a new file in the current folder.")));
             } else if (i > 8 && folders.size() > i-9) {
                 content.add(folders.get(i-9));
             } else {
@@ -80,6 +90,12 @@ public class FileGui implements Listener {
     }
 
     public static void openFile(Player player) {
+        if (!player.hasPermission("fileman.file.read")) {
+            fileData.remove(player.getUniqueId());
+            position.remove(player.getName());
+            return;
+        }
+
         FileData data =  fileData.get(player.getUniqueId());
 
         int lineNumber = data.getCurrentLine();
@@ -91,8 +107,9 @@ public class FileGui implements Listener {
         if (!file.isFile()) return;
 
         ArrayList<String> lines = new ArrayList<>();
+        BufferedReader fileReader = null;
         try {
-            BufferedReader fileReader = new BufferedReader(new FileReader(file));
+           fileReader = new BufferedReader(new FileReader(file));
 
             String text;
             while ((text = fileReader.readLine()) != null)
@@ -100,7 +117,18 @@ public class FileGui implements Listener {
 
         } catch (IOException e) {
             log(e, player, Level.WARNING, "There was an error trying to read \""+position.get(player.getName()).getRelativePath()+"\".");
+            try {
+                fileReader.close();
+            } catch (Exception ex) {
+                log(e, player, Level.WARNING, "There was an error closing the file \""+file.getName()+"\". This file will not be able to be used until the server is restarted.");
+                return;
+            }
             return;
+        }
+        try {
+            fileReader.close();
+        } catch (IOException e) {
+            log(e, player, Level.WARNING, "There was an error closing the file \""+file.getName()+"\". This file will not be able to be used until the server is restarted.");
         }
 
         if (lineNumber > lines.size()) lineNumber = lines.size();
@@ -182,12 +210,20 @@ public class FileGui implements Listener {
             }
         }
 
+
+
         fileData.put(player.getUniqueId(), data);
         gui.setContents(content.toArray(new ItemStack[0]));
         player.openInventory(gui);
     }
 
     public static void editFile(Player player, ItemStack editedText) {
+        if (!player.hasPermission("fileman.file.edit")) {
+            fileData.remove(player.getUniqueId());
+            position.remove(player.getName());
+            return;
+        }
+
         if (editedText.getItemMeta() == null) return;
         PathHolder pathHolder = position.get(player.getName());
         Integer line = editedText.getItemMeta().getPersistentDataContainer().get(new NamespacedKey(JavaPlugin.getPlugin(FileManager.class), "line"), PersistentDataType.INTEGER);
@@ -222,13 +258,27 @@ public class FileGui implements Listener {
         HumanEntity player = e.getWhoClicked();
         String inventoryTitle = player.getOpenInventory().getTitle();
         if (!position.containsKey(player.getName())) return;
+        //TODO: think of something better
         if (inventoryTitle.equals(ChatColor.BLUE+"~"+position.get(player.getName()).getRelativePath().substring(0, position.get(player.getName()).getRelativePath().length()-1)+ChatColor.GOLD+" $")
                 || inventoryTitle.equals(ChatColor.BLUE+"~"+position.get(player.getName()).getRelativePath()+ChatColor.GOLD+" $")
-                || inventoryTitle.startsWith("Min: 1, Max:") || inventoryTitle.startsWith("Search:") || inventoryTitle.startsWith("File editor:"))
+                || inventoryTitle.startsWith("Min: 1, Max:") || inventoryTitle.startsWith("Search:") || inventoryTitle.startsWith("File editor:")
+                || inventoryTitle.equals("File or Folder") || inventoryTitle.startsWith("Name:"))
             e.setCancelled(true);
         else {
             position.remove(player.getName());
             fileData.remove(player.getUniqueId());
+        }
+    }
+
+    @EventHandler
+    public void upFolder(InventoryClickEvent e) {
+        if (e.getWhoClicked() instanceof Player player) {
+            if (e.getCurrentItem() == null || e.getCurrentItem().getItemMeta() == null || e.getCurrentItem().getType() != Material.ARROW || !position.containsKey(player.getName())) return;
+            String identifier = e.getCurrentItem().getItemMeta().getPersistentDataContainer().get(new NamespacedKey(JavaPlugin.getPlugin(FileManager.class), "identifier"), PersistentDataType.STRING);
+            if (identifier != null) return;
+            PathHolder pathHolder = position.get(player.getName());
+            pathHolder.setCurrentPath(Path.of(pathHolder.getCurrentPath()).getParent().toString());
+            openFolder(player);
         }
     }
 
@@ -243,6 +293,74 @@ public class FileGui implements Listener {
     }
 
     @EventHandler
+    public void createMenu(InventoryClickEvent e) {
+        if (e.getWhoClicked() instanceof Player player) {
+            if (!player.hasPermission("fileman.file.create")) return;
+            if (e.getCurrentItem() == null || e.getCurrentItem().getItemMeta() == null || e.getCurrentItem().getType() != Material.GREEN_CONCRETE || !position.containsKey(player.getName())) return;
+            player.closeInventory();
+            Inventory gui = Bukkit.createInventory(player, InventoryType.DROPPER, "File or Folder");
+
+            ArrayList<ItemStack> content = new ArrayList<>();
+            for (int i = 0; i <= 8; i++) {
+                if (i == 3) {
+                    content.add(itemProperties(new ItemStack(Material.WHITE_WOOL), "File", null));
+                } else if (i == 5) {
+                    content.add(itemProperties(new ItemStack(Material.YELLOW_WOOL), ChatColor.YELLOW + "Folder", null));
+                } else {
+                    content.add(new ItemStack(Material.LIGHT_GRAY_STAINED_GLASS_PANE));
+                }
+            }
+
+            gui.setContents(content.toArray(new ItemStack[0]));
+            player.openInventory(gui);
+        }
+    }
+
+    @EventHandler
+    public void createFile(InventoryClickEvent e) {
+        if (e.getWhoClicked() instanceof Player player) {
+            if (!player.hasPermission("fileman.file.create")) return;
+            if (e.getCurrentItem() == null || e.getCurrentItem().getItemMeta() == null || !position.containsKey(player.getName()) || !player.getOpenInventory().getTitle().equals("File or Folder") || e.getCurrentItem().getType() == Material.LIGHT_GRAY_STAINED_GLASS_PANE || e.getCurrentItem().getType() == Material.GREEN_CONCRETE) return;
+            player.closeInventory();
+
+            ItemStack item = e.getCurrentItem();
+            ItemMeta meta = item.getItemMeta();
+            meta.getDisplayName().replaceAll(String.valueOf(ChatColor.YELLOW), "");
+            item.setItemMeta(meta);
+
+            new AnvilGUI.Builder()
+                    .plugin(JavaPlugin.getPlugin(FileManager.class))
+                    .preventClose()
+                    .title("Name:")
+                    .itemLeft(item)
+                    .onClick((slot, stateSnapshot) -> {
+                        if (slot == AnvilGUI.Slot.OUTPUT) {
+                            return List.of(AnvilGUI.ResponseAction.close());
+                        }
+                        return Collections.emptyList();
+                    })
+                    .onClose(stateSnapshot -> {
+                        if (stateSnapshot.getOutputItem().getItemMeta() == null) return;
+                        try {
+                            if (stateSnapshot.getOutputItem().getType() == Material.YELLOW_WOOL) {
+                                Files.createDirectory(Path.of(position.get(stateSnapshot.getPlayer().getName()).getCurrentPath() + File.separator + stateSnapshot.getOutputItem().getItemMeta().getDisplayName()));
+                            } else if (stateSnapshot.getOutputItem().getType() == Material.WHITE_WOOL) {
+                                Files.createFile(Path.of(position.get(stateSnapshot.getPlayer().getName()).getCurrentPath() + File.separator + stateSnapshot.getOutputItem().getItemMeta().getDisplayName()));
+                            } else {
+                                return;
+                            }
+                        } catch (FileAlreadyExistsException ex) {
+                            log(ex, stateSnapshot.getPlayer(), Level.WARNING, "There is already a file/folder with that name.");
+                        } catch (IOException ex) {
+                            log(ex, stateSnapshot.getPlayer(), Level.WARNING, "There was an error creating that file/folder.");
+                        }
+                        openFolder(stateSnapshot.getPlayer());
+                    })
+                    .open(player);
+        }
+    }
+
+    @EventHandler
     public void useFolder(InventoryClickEvent e) {
         if (e.getWhoClicked() instanceof Player player) {
             if (e.getCurrentItem() == null || e.getCurrentItem().getItemMeta() == null || e.getCurrentItem().getItemMeta().getLore() == null || !position.containsKey(player.getName())) return;
@@ -251,18 +369,6 @@ public class FileGui implements Listener {
                 pathHolder.setCurrentPath(pathHolder.getCurrentPath() + File.separator + e.getCurrentItem().getItemMeta().getPersistentDataContainer().get(new NamespacedKey(JavaPlugin.getPlugin(FileManager.class), "identifier"), PersistentDataType.STRING));
                 openFolder(player);
             }
-        }
-    }
-
-    @EventHandler
-    public void upFolder(InventoryClickEvent e) {
-        if (e.getWhoClicked() instanceof Player player) {
-            if (e.getCurrentItem() == null || e.getCurrentItem().getItemMeta() == null || e.getCurrentItem().getType() != Material.ARROW || !position.containsKey(player.getName())) return;
-            String identifier = e.getCurrentItem().getItemMeta().getPersistentDataContainer().get(new NamespacedKey(JavaPlugin.getPlugin(FileManager.class), "identifier"), PersistentDataType.STRING);
-            if (identifier != null) return;
-            PathHolder pathHolder = position.get(player.getName());
-            pathHolder.setCurrentPath(Path.of(pathHolder.getCurrentPath()).getParent().toString());
-            openFolder(player);
         }
     }